+++ libs/auth/src/angular/input-password/input-password.component.html
@@ -1,7 +1,7 @@
 <form [formGroup]="formGroup" [bitSubmit]="submit">
   <auth-password-callout
-    *ngIf="masterPasswordPolicy"
-    [policy]="masterPasswordPolicy"
+    *ngIf="masterPasswordPolicyOptions"
+    [policy]="masterPasswordPolicyOptions"
   ></auth-password-callout>
 
   <div class="tw-mb-6">
