+++ libs/auth/src/angular/input-password/input-password.mdx
@@ -22,8 +22,8 @@ the parent component to act on those values as needed.
   `InputPasswordComponent` can create a master key.
 - `buttonText` (optional) - an `i18n` translated string that can be used as button text (default
   text is "Set master password").
-- `orgId` (optional) - used to retreive and enforce the master password policy requirements for an
-  org.
+- `masterPasswordPolicyOptions` (optional) - used to display and enforce master password policy
+  requirements.
 
 <br />
