+++ libs/auth/src/angular/input-password/input-password.stories.ts
@@ -2,11 +2,9 @@ import { importProvidersFrom } from "@angular/core";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { action } from "@storybook/addon-actions";
 import { Meta, StoryObj, applicationConfig } from "@storybook/angular";
-import { of } from "rxjs";
 import { ZXCVBNResult } from "zxcvbn";
 
 import { AuditService } from "@bitwarden/common/abstractions/audit.service";
-import { PolicyApiServiceAbstraction } from "@bitwarden/common/admin-console/abstractions/policy/policy-api.service.abstraction";
 import { PolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
 import { MasterPasswordPolicyOptions } from "@bitwarden/common/admin-console/models/domain/master-password-policy-options";
 import { CryptoService } from "@bitwarden/common/platform/abstractions/crypto.service";
@@ -17,22 +15,10 @@ import { PreloadedEnglishI18nModule } from "../../../../../apps/web/src/app/core
 
 import { InputPasswordComponent } from "./input-password.component";
 
-const mockMasterPasswordPolicyOptions = {
-  minComplexity: 4,
-  minLength: 14,
-  requireUpper: true,
-  requireLower: true,
-  requireNumbers: true,
-  requireSpecial: true,
-} as MasterPasswordPolicyOptions;
-
 export default {
   title: "Auth/Input Password",
   component: InputPasswordComponent,
-} as Meta;
-
-const decorators = (options: { hasPolicy?: boolean }) => {
-  return [
+  decorators: [
     applicationConfig({
       providers: [
         importProvidersFrom(PreloadedEnglishI18nModule),
@@ -57,17 +43,8 @@ const decorators = (options: { hasPolicy?: boolean }) => {
           } as Partial<DialogService>,
         },
         {
-          provide: PolicyApiServiceAbstraction,
-          useValue: {
-            getMasterPasswordPolicyOptsForOrgUser: () =>
-              options.hasPolicy ? mockMasterPasswordPolicyOptions : null,
-          } as Partial<PolicyService>,
-        },
-        {
           provide: PolicyService,
           useValue: {
-            masterPasswordPolicyOptions$: () =>
-              options.hasPolicy ? of(mockMasterPasswordPolicyOptions) : null,
             evaluateMasterPassword: (score) => {
               if (score < 4) {
                 return false;
@@ -81,7 +58,6 @@ const decorators = (options: { hasPolicy?: boolean }) => {
           useValue: {
             getPasswordStrength: (password) => {
               let score = 0;
-
               if (password.length === 0) {
                 score = null;
               } else if (password.length <= 4) {
@@ -93,7 +69,6 @@ const decorators = (options: { hasPolicy?: boolean }) => {
               } else {
                 score = 4;
               }
-
               return { score } as ZXCVBNResult;
             },
           } as Partial<PasswordStrengthServiceAbstraction>,
@@ -106,8 +81,18 @@ const decorators = (options: { hasPolicy?: boolean }) => {
         },
       ],
     }),
-  ];
-};
+  ],
+  args: {
+    masterPasswordPolicyOptions: {
+      minComplexity: 4,
+      minLength: 14,
+      requireUpper: true,
+      requireLower: true,
+      requireNumbers: true,
+      requireSpecial: true,
+    } as MasterPasswordPolicyOptions,
+  },
+} as Meta;
 
 type Story = StoryObj<InputPasswordComponent>;
 
@@ -118,19 +103,13 @@ export const Default: Story = {
       <auth-input-password></auth-input-password>
     `,
   }),
-  decorators: decorators({
-    hasPolicy: false,
-  }),
 };
 
 export const WithPolicy: Story = {
   render: (args) => ({
     props: args,
     template: `
-      <auth-input-password></auth-input-password>
+      <auth-input-password [masterPasswordPolicyOptions]="masterPasswordPolicyOptions"></auth-input-password>
     `,
-  }),
-  decorators: decorators({
-    hasPolicy: true,
   }),
 };
