+++ libs/auth/src/angular/input-password/password-input-result.ts
@@ -1,0 +1,9 @@
+import { PBKDF2KdfConfig } from "@bitwarden/common/auth/models/domain/kdf-config";
+import { MasterKey } from "@bitwarden/common/types/key";
+
+export interface PasswordInputResult {
+  masterKey: MasterKey;
+  masterKeyHash: string;
+  kdfConfig: PBKDF2KdfConfig;
+  hint: string;
+}
